uavcan_v1: Reorganize uavcan params and Subscriber
